% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_word_containing_my_word.R
\name{get_random_word_containing_my_word}
\alias{get_random_word_containing_my_word}
\title{Random word containing provided word}
\usage{
get_random_word_containing_my_word(myword)
}
\arguments{
\item{myword}{This is a string.}
}
\description{
This function takes a word as input and outputs a random word that contains the given word.
}
\examples{
get_random_word_containing_my_word("evolution")
}
